New Background, existing to GameEnvBackgroud, support for Parallax
